Fix linter errors, create a new module for the books
